add scoring function in handlequiz function
